+++ libs/operation_mode/CMakeLists.txt
@@ -17,7 +17,6 @@ set_target_properties(liboperation_mode PROPERTIES PREFIX "")
 
 find_package(Boost REQUIRED)
 find_package(unofficial-sdbusplus CONFIG REQUIRED)
-find_package(glaze CONFIG REQUIRED)
 find_package(magic_enum CONFIG REQUIRED)
 find_package(mp CONFIG REQUIRED)
 target_link_libraries(liboperation_mode
@@ -25,7 +24,7 @@ target_link_libraries(liboperation_mode
     tfc::base
     tfc::logger
     tfc::stx
-    glaze::glaze
+    tfc::dbus_util
     magic_enum::magic_enum
     Boost::mp
   PRIVATE
