+++ libs/stx/CMakeLists.txt
@@ -9,9 +9,3 @@ target_include_directories(stx
     $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/inc/public>
     $<INSTALL_INTERFACE:inc/public>
 )
-
-# Those links should be minimized, consider splitting utils and stx if utils needs further dependencies
-target_link_libraries(stx
-  INTERFACE
-    tfc::configure_options
-)
